feat: ComponentRegister 구현
- ShoppingApplication 내 인스턴스 등록 로직을 ComponentRegister에 구현
